requirements.txt
pandas==1.5.3
numpy==1.23.5
plotly==5.9.0
streamlit==1.12.0
Pillow==9.0.1
folium==0.12.1
streamlit-folium==0.8.0